pip_requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio-sdk==0.3.7
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.9.0
astor==0.8.1
asyncpg==0.30.0
attrs==25.3.0
av==15.1.0
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.45
beautifulsoup4==4.13.5
blake3==1.0.5
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
cbor2==5.7.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.2
cloudpathlib==0.22.0
cloudpickle==3.1.1
cohere==5.18.0
coloredlogs==15.0.1
colorlog==6.9.0
compressed-tensors==0.11.0
confection==0.1.5
cryptography==45.0.7
cssselect==1.3.0
cssutils==2.11.1
csvw==3.6.0
ctranslate2==4.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dacite==1.9.2
dataclasses-json==0.6.7
datasets==4.1.0
deprecation==2.1.0
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
docopt==0.6.2
docx2txt==0.9
ecdsa==0.19.1
einops==0.8.1
email-validator==2.3.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.12.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastavro==1.12.0
faster-whisper==1.2.0
filelock==3.18.0
filetype==1.2.0
FlashRank==0.2.10
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
future==1.0.0
gguf==0.17.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.36.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.10
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imagesize==1.4.1
interegular==0.3.3
invoke==2.2.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kokoro==0.9.4
langchain==0.3.27
langchain-anthropic==0.3.20
langchain-cohere==0.4.6
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-genai==2.1.10
langchain-openai==0.3.28
langchain-postgres==0.0.15
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.28
language-tags==1.2.0
language_data==1.3.0
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
misaki==0.9.4
mistral_common==1.8.5
mistralai==1.9.10
mlx==0.29.1
mlx-lm==0.27.1
mlx-metal==0.29.1
modelscope==1.29.2
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
olefile==0.47
onnxruntime==1.22.1
openai==1.107.2
openai-harmony==0.0.4
openai-whisper==20250625
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.1
ormsgpack==1.10.0
outlines_core==0.2.11
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.1
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
pgvector==0.3.6
phonemizer-fork==3.3.2
pillow==11.3.0
postgrest==1.1.1
premailer==3.10.0
preshed==3.0.10
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
psycopg==3.2.9
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.4
pyparsing==3.2.4
pypdf==6.0.0
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-jose==3.5.0
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
rank-llm==0.25.7
RapidFuzz==3.14.1
rdflib==7.2.1
realtime==2.6.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
segments==2.3.0
sentence-transformers==5.1.0
sentencepiece==0.2.1
sentry-sdk==2.38.0
setproctitle==1.3.7
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
srsly==2.5.1
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
types-PyYAML==6.0.12.20250822
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
unstructured==0.18.14
unstructured-client==0.42.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.2
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
xgrammar==0.1.23
xlsxwriter==3.2.8
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
